+++ presentation/src/main/java/com/hihihihi/gureumpage/ui/mypage/MypageViewModel.kt
@@ -186,113 +186,6 @@ class MypageViewModel @Inject constructor(
             .onFailure { e -> _uiState.update { it.copy(error = e.message) } }
     }
 
-    // withdrawUser 함수 개선
-    fun withdrawUser() = viewModelScope.launch {
-        _uiState.update { it.copy(loading = true, error = null) }
-
-        try {
-            val currentUser = FirebaseAuth.getInstance().currentUser
-            if (currentUser == null) {
-                _uiState.update { it.copy(loading = false, error = "로그인된 사용자가 없습니다") }
-                return@launch
-            }
-
-            val providerId = getProviderFromFirestore(currentUser.uid)
-            Log.d("WITHDRAWAL", "탈퇴 시작 - Provider: $providerId")
-
-            // 1. 먼저 소셜 플랫폼에서 연결 해제
-            when (providerId) {
-                "kakao" -> {
-                    Log.d("WITHDRAWAL", "카카오 연결 해제 시작")
-                    unlinkKakaoAccount()
-                }
-                "naver" -> {
-                    Log.d("WITHDRAWAL", "네이버 연결 해제 시작")
-                    unlinkNaverAccount()
-                }
-                "google.com" -> {
-                    Log.d("WITHDRAWAL", "구글 - Firebase에서 자동 처리")
-                }
-                else -> {
-                    Log.d("WITHDRAWAL", "알 수 없는 provider: $providerId")
-                }
-            }
-
-            // 2. Firebase Functions 호출 (Firestore + Auth 삭제)
-            Log.d("WITHDRAWAL", "Firebase Functions 호출 시작")
-            val functions = Firebase.functions
-            val deleteAccount = functions.getHttpsCallable("deleteUserAccount")
-            val result = deleteAccount.call().await()
-            Log.d("WITHDRAWAL", "Functions 호출 성공: $result")
-
-            // 3. 상태 초기화 및 로그아웃 이벤트 발생
-            _uiState.value = MyPageUiState(loading = false)
-            _logoutEvent.tryEmit(Unit)
-
-            Log.d("WITHDRAWAL", "탈퇴 완료")
-
-        } catch (e: Exception) {
-            Log.e("WITHDRAWAL", "탈퇴 실패", e)
-
-            val errorMessage = when {
-                e.message?.contains("unauthenticated") == true -> "인증이 필요합니다"
-                e.message?.contains("not-found") == true -> "사용자를 찾을 수 없습니다"
-                e.message?.contains("permission-denied") == true -> "권한이 없습니다"
-                else -> "탈퇴 처리 중 오류가 발생했습니다: ${e.message}"
-            }
-
-            _uiState.update { it.copy(loading = false, error = errorMessage) }
-        }
-    }
-
-    // Firestore에서 provider 가져오는 함수
-    private suspend fun getProviderFromFirestore(uid: String): String? {
-        return try {
-            val firestore = FirebaseFirestore.getInstance()
-            val userDoc = firestore.collection("users").document(uid).get().await()
-            val provider = userDoc.getString("provider")
-            Log.d("WITHDRAWAL", "Firestore에서 가져온 provider: $provider")
-            provider
-        } catch (e: Exception) {
-            Log.e("WITHDRAWAL", "Provider 가져오기 실패", e)
-            null
-        }
-    }
-
-    private suspend fun unlinkKakaoAccount() {
-        return suspendCoroutine { continuation ->
-            UserApiClient.instance.unlink { error ->
-                if (error != null) {
-                    Log.e("KAKAO", "카카오 연결 해제 실패", error)
-                    continuation.resumeWithException(error)
-                } else {
-                    Log.i("KAKAO", "카카오 연결 해제 성공")
-                    continuation.resume(Unit)
-                }
-            }
-        }
-    }
-
-    private suspend fun unlinkNaverAccount() {
-        return suspendCoroutine { continuation ->
-            NidOAuthLogin().callDeleteTokenApi(object : OAuthLoginCallback {
-                override fun onSuccess() {
-                    Log.d("NAVER", "네이버 연결 해제 성공")
-                    continuation.resume(Unit)
-                }
-
-                override fun onFailure(httpStatus: Int, message: String) {
-                    Log.e("NAVER", "네이버 연결 해제 실패: $message")
-                    continuation.resumeWithException(Exception(message))
-                }
-
-                override fun onError(errorCode: Int, message: String) {
-                    Log.e("NAVER", "네이버 연결 해제 에러: $message")
-                    continuation.resumeWithException(Exception(message))
-                }
-            })
-        }
-    }
 
 
 }
