+++ presentation/src/main/java/com/hihihihi/gureumpage/ui/withdraw/WithdrawUiState.kt
@@ -1,0 +1,6 @@
+package com.hihihihi.gureumpage.ui.withdraw
+
+data class WithdrawUiState(
+    val loading: Boolean = false,
+    val error: String? = null
+)
